refactor: 배팅금 입력 클래스 내부 메서드 분리 및 클래스 이름 변경
