--- src/main/java/view/BattingMoneyInputView.java
@@ -1,27 +1,0 @@
-package view;
-
-import java.util.List;
-
-public class BattingMoneyInputView {
-    public static List<Integer> getMoney(List<String> playersName) {
-        return playersName.stream()
-                .peek(playerName -> Console.printf("%s의 배팅 금액은?%n", playerName))
-                .map(playerName -> {
-                    String inputFromConsole = Console.getInputFromConsole();
-                    validateInput(inputFromConsole);
-                    return Integer.parseInt(inputFromConsole);
-                }).toList();
-    }
-
-    private static void validateInput(String input) {
-        int rawNumber;
-        try {
-            rawNumber = Integer.parseInt(input);
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("입력 형식이 올바르지 않습니다.");
-        }
-        if (rawNumber <= 0) {
-            throw new IllegalArgumentException("입력 형식이 올바르지 않습니다.");
-        }
-    }
-}
